Updated code for Pyslowfast 1.0, Pytorch 1.6 and cuda 10.2
