+++ configs/SLOWFAST_8x8_R50.yaml
@@ -1,11 +1,11 @@
 TRAIN:
   ENABLE: False
   DATASET: videoset
-  CHECKPOINT_FILE_PATH: "./checkpoints/SLOWFAST_8x8_R50.pkl"
+  CHECKPOINT_FILE_PATH: "path/to/checkpoint"
   CHECKPOINT_TYPE: caffe2
 DATA:
-  PATH_TO_DATA_DIR: "/media/data/tridiv/epic_kitchens/EPIC_KITCHENS_2018"
-  PATH_PREFIX: "frames_rgb_flow/rgb"
+  PATH_TO_DATA_DIR: "/path/to/dataset"
+  PATH_PREFIX: ""
   NUM_FRAMES: 32
   SAMPLING_RATE: 2
   TRAIN_JITTER_SCALES: [256, 320]
@@ -13,8 +13,8 @@ DATA:
   TEST_CROP_SIZE: 256
   INPUT_CHANNEL_NUM: [3, 3]
   IN_FPS: 60
-  OUT_FPS: 30
-  READ_VID_FILE: False
+  OUT_FPS: 60
+  READ_VID_FILE: True
   VID_FILE_EXT: ".MP4"
   IMG_FILE_EXT: ".jpg"
   IMG_FILE_FORMAT: "frame_{:010d}.jpg"
@@ -32,6 +32,8 @@ RESNET:
   TRANS_FUNC: bottleneck_transform
   STRIDE_1X1: False
   NUM_BLOCK_TEMP_KERNEL: [[3, 3], [4, 4], [6, 6], [3, 3]]
+  SPATIAL_STRIDES: [[1, 1], [2, 2], [2, 2], [2, 2]]
+  SPATIAL_DILATIONS: [[1, 1], [1, 1], [1, 1], [1, 1]]
 NONLOCAL:
   LOCATION: [[[], []], [[], []], [[], []], [[], []]]
   GROUP: [[1, 1], [1, 1], [1, 1], [1, 1]]
@@ -39,30 +41,31 @@ NONLOCAL:
 BN:
   USE_PRECISE_STATS: True
   NUM_BATCHES_PRECISE: 200
-  MOMENTUM: 0.1
-  WEIGHT_DECAY: 0.0
 SOLVER:
   BASE_LR: 0.1
   LR_POLICY: cosine
   MAX_EPOCH: 196
   MOMENTUM: 0.9
   WEIGHT_DECAY: 1e-4
-  WARMUP_EPOCHS: 37
-  WARMUP_START_LR: 0.001
+  WARMUP_EPOCHS: 34.0
+  WARMUP_START_LR: 0.01
   OPTIMIZING_METHOD: sgd
 MODEL:
   NUM_CLASSES: 400
   ARCH: slowfast
+  MODEL_NAME: SlowFast
   LOSS_FUNC: cross_entropy
   DROPOUT_RATE: 0.5
 TEST:
   ENABLE: True
   DATASET: videoset
-  BATCH_SIZE: 3
+  BATCH_SIZE: 1
 DATA_LOADER:
   NUM_WORKERS: 8
   PIN_MEMORY: True
+DETECTION:
+  ENABLE: False
 NUM_GPUS: 1
 NUM_SHARDS: 1
 RNG_SEED: 0
-OUTPUT_DIR: "/media/data/tridiv/epic_features"
+OUTPUT_DIR: "path/to/output/features"
