+++ configs/custom_config.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 
 import slowfast.config.defaults as defcfg
+import slowfast.utils.checkpoint as cu
 
 
 # -----------------------------------------------------------------------------
@@ -34,3 +35,34 @@ def get_cfg():
     Get a copy of the default config.
     """
     return defcfg._assert_and_infer_cfg(defcfg._C.clone())
+
+
+def load_config(args):
+    """
+    Given the arguemnts, load and initialize the configs.
+    Args:
+        args (argument): arguments includes `shard_id`, `num_shards`,
+            `init_method`, `cfg_file`, and `opts`.
+    """
+    # Setup cfg.
+    cfg = get_cfg()
+    # Load config from cfg.
+    if args.cfg_file is not None:
+        cfg.merge_from_file(args.cfg_file)
+    # Load config from command line, overwrite config from opts.
+    if args.opts is not None:
+        cfg.merge_from_list(args.opts)
+
+    # Inherit parameters from args.
+    if hasattr(args, "num_shards") and hasattr(args, "shard_id"):
+        cfg.NUM_SHARDS = args.num_shards
+        cfg.SHARD_ID = args.shard_id
+    if hasattr(args, "rng_seed"):
+        cfg.RNG_SEED = args.rng_seed
+    if hasattr(args, "output_dir"):
+        cfg.OUTPUT_DIR = args.output_dir
+
+    # Create the checkpoint dir.
+    # cu.make_checkpoint_dir(cfg.OUTPUT_DIR)
+
+    return cfg
