--- datasets/loader.py
@@ -1,76 +1,0 @@
-#!/usr/bin/env python3
-# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
-
-"""Data loader."""
-
-import torch
-from torch.utils.data.distributed import DistributedSampler
-from torch.utils.data.sampler import RandomSampler
-
-from slowfast.datasets.kinetics import Kinetics
-from datasets.videoset import VideoSet
-
-# Supported datasets.
-_DATASET_CATALOG = {"kinetics": Kinetics}
-
-
-def construct_loader(cfg, split):
-    """
-    Constructs the data loader for the given dataset.
-    Args:
-        cfg (CfgNode): configs. Details can be found in
-            slowfast/config/defaults.py
-        split (str): the split of the data loader. Options include `train`,
-            `val`, and `test`.
-    """
-    assert split in ["train", "val", "test"]
-    if split in ["train"]:
-        dataset_name = cfg.TRAIN.DATASET
-        batch_size = int(cfg.TRAIN.BATCH_SIZE / cfg.NUM_GPUS)
-        shuffle = True
-        drop_last = True
-    elif split in ["val"]:
-        dataset_name = cfg.TRAIN.DATASET
-        batch_size = int(cfg.TRAIN.BATCH_SIZE / cfg.NUM_GPUS)
-        shuffle = False
-        drop_last = False
-    elif split in ["test"]:
-        dataset_name = cfg.TEST.DATASET
-        batch_size = int(cfg.TEST.BATCH_SIZE / cfg.NUM_GPUS)
-        shuffle = False
-        drop_last = False
-    assert (
-        dataset_name in _DATASET_CATALOG.keys()
-    ), "Dataset '{}' is not supported".format(dataset_name)
-
-    # Construct the dataset
-    dataset = _DATASET_CATALOG[dataset_name](cfg, split)
-    # Create a sampler for multi-process training
-    sampler = DistributedSampler(dataset) if cfg.NUM_GPUS > 1 else None
-    # Create a loader
-    loader = torch.utils.data.DataLoader(
-        dataset,
-        batch_size=batch_size,
-        shuffle=(False if sampler else shuffle),
-        sampler=sampler,
-        num_workers=cfg.DATA_LOADER.NUM_WORKERS,
-        pin_memory=cfg.DATA_LOADER.PIN_MEMORY,
-        drop_last=drop_last,
-    )
-    return loader
-
-
-def shuffle_dataset(loader, cur_epoch):
-    """"
-    Shuffles the data.
-    Args:
-        loader (loader): data loader to perform shuffle.
-        cur_epoch (int): number of the current epoch.
-    """
-    assert isinstance(
-        loader.sampler, (RandomSampler, DistributedSampler)
-    ), "Sampler type '{}' not supported".format(type(loader.sampler))
-    # RandomSampler handles shuffling automatically
-    if isinstance(loader.sampler, DistributedSampler):
-        # DistributedSampler shuffles data based on epoch
-        loader.sampler.set_epoch(cur_epoch)
