+++ models/__init__.py
@@ -1,2 +1,3 @@
-#!/usr/bin/env python3
-# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+from .video_model_builder import ResNetFeat, SlowFastFeat
+from .head_helper import ResNetBasicHead
+from .build import build_model
