+++ models/head_helper.py
@@ -1,74 +1,13 @@
 #!/usr/bin/env python3
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
 
-"""ResNe(X)t Head helper."""
-
 import torch
-import torch.nn as nn
+
+from slowfast.models.head_helper import ResNetBasicHead
 
 
-class ResNetBasicHead(nn.Module):
-    """
-    ResNe(X)t 3D head.
-    This layer performs a fully-connected projection during training, when the
-    input size is 1x1x1. It performs a convolutional projection during testing
-    when the input size is larger than 1x1x1. If the inputs are from multiple
-    different pathways, the inputs will be concatenated after pooling.
-    """
-
-    def __init__(
-        self,
-        dim_in,
-        num_classes,
-        pool_size,
-        dropout_rate=0.0,
-        act_func="softmax",
-    ):
-        """
-        The `__init__` method of any subclass should also contain these
-            arguments.
-        ResNetBasicHead takes p pathways as input where p in [1, infty].
-
-        Args:
-            dim_in (list): the list of channel dimensions of the p inputs to the
-                ResNetHead.
-            num_classes (int): the channel dimensions of the p outputs to the
-                ResNetHead.
-            pool_size (list): the list of kernel sizes of p spatial temporal
-                poolings, temporal pool kernel size, spatial pool kernel size,
-                spatial pool kernel size in order.
-            dropout_rate (float): dropout rate. If equal to 0.0, perform no
-                dropout.
-            act_func (string): activation function to use. 'softmax': applies
-                softmax on the output. 'sigmoid': applies sigmoid on the output.
-        """
-        super(ResNetBasicHead, self).__init__()
-        assert (
-            len({len(pool_size), len(dim_in)}) == 1
-        ), "pathway dimensions are not consistent."
-        self.num_pathways = len(pool_size)
-
-        for pathway in range(self.num_pathways):
-            avg_pool = nn.AvgPool3d(pool_size[pathway], stride=1)
-            self.add_module("pathway{}_avgpool".format(pathway), avg_pool)
-
-        if dropout_rate > 0.0:
-            self.dropout = nn.Dropout(dropout_rate)
-        # Perform FC in a fully convolutional manner. The FC layer will be
-        # initialized with a different std comparing to convolutional layers.
-        self.projection = nn.Linear(sum(dim_in), num_classes, bias=True)
-
-        # Softmax for evaluation and testing.
-        if act_func == "softmax":
-            self.act = nn.Softmax(dim=4)
-        elif act_func == "sigmoid":
-            self.act = nn.Sigmoid()
-        else:
-            raise NotImplementedError(
-                "{} is not supported as an activation"
-                "function.".format(act_func)
-            )
-
+class ResNetBasicHead(ResNetBasicHead):
+    # Overwrite function to return features
     def forward(self, inputs):
         assert (
             len(inputs) == self.num_pathways
@@ -85,7 +24,7 @@ class ResNetBasicHead(nn.Module):
         feat = x.clone().detach()
         # flatten the features tensor
         feat = feat.mean(3).mean(2).reshape(feat.shape[0], -1)
-        
+
         # Perform dropout.
         if hasattr(self, "dropout"):
             x = self.dropout(x)
