--- models/model_builder.py
@@ -1,45 +1,0 @@
-#!/usr/bin/env python3
-# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
-
-"""Model construction functions."""
-
-import torch
-
-from models.video_model_builder import ResNetModel, SlowFastModel
-
-# Supported model types
-_MODEL_TYPES = {
-    "slowfast": SlowFastModel,
-    "slowonly": ResNetModel,
-    "c2d": ResNetModel,
-    "i3d": ResNetModel,
-}
-
-
-def build_model(cfg):
-    """
-    Builds the video model.
-    Args:
-        cfg (configs): configs that contains the hyper-parameters to build the
-        backbone. Details can be seen in slowfast/config/defaults.py.
-    """
-    assert (
-        cfg.MODEL.ARCH in _MODEL_TYPES.keys()
-    ), "Model type '{}' not supported".format(cfg.MODEL.ARCH)
-    assert (
-        cfg.NUM_GPUS <= torch.cuda.device_count()
-    ), "Cannot use more GPU devices than available"
-
-    # Construct the model
-    model = _MODEL_TYPES[cfg.MODEL.ARCH](cfg)
-    # Determine the GPU used by the current process
-    cur_device = torch.cuda.current_device()
-    # Transfer the model to the current GPU device
-    model = model.cuda(device=cur_device)
-    # Use multi-process data parallel model in the multi-gpu setting
-    if cfg.NUM_GPUS > 1:
-        # Make model replica operate on the current device
-        model = torch.nn.parallel.DistributedDataParallel(
-            module=model, device_ids=[cur_device], output_device=cur_device
-        )
-    return model
